refactor: extract createCircleBuffer to shared geo utility
Deduplicate the identical Haversine circle polygon generator that was
defined inline in both ConsultationMap and NotificationMapDialog.
