+++ src/components/admin/NotificationMapDialog.tsx
@@ -4,6 +4,7 @@ import { useState, useEffect, useMemo } from 'react';
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription } from '@/components/ui/dialog';
 import { Loader2 } from 'lucide-react';
 import Map, { MapFeature } from '@/components/map/map';
+import { createCircleBuffer } from '@/lib/geo/buffer';
 
 interface SubjectLocation {
     id: string;
@@ -29,38 +30,6 @@ interface NotificationMapDialogProps {
     meetingId: string;
     cityId: string;
     meetingName: string;
-}
-
-// Helper function to create a circular polygon buffer around a point
-function createCircleBuffer(center: [number, number], radiusInMeters: number): GeoJSON.Polygon {
-    const earthRadius = 6371000; // Earth's radius in meters
-    const lat = center[1] * Math.PI / 180; // Convert to radians
-    const lng = center[0] * Math.PI / 180;
-
-    const points: [number, number][] = [];
-    const numPoints = 64; // Number of points to create the circle
-
-    for (let i = 0; i < numPoints; i++) {
-        const angle = (i * 360 / numPoints) * Math.PI / 180;
-
-        // Calculate offset in radians
-        const dLat = radiusInMeters * Math.cos(angle) / earthRadius;
-        const dLng = radiusInMeters * Math.sin(angle) / (earthRadius * Math.cos(lat));
-
-        // Convert back to degrees
-        const newLat = (lat + dLat) * 180 / Math.PI;
-        const newLng = (lng + dLng) * 180 / Math.PI;
-
-        points.push([newLng, newLat]);
-    }
-
-    // Close the polygon by adding the first point at the end
-    points.push(points[0]);
-
-    return {
-        type: 'Polygon',
-        coordinates: [points]
-    };
 }
 
 export function NotificationMapDialog({
