+++ src/components/consultations/ConsultationMap.tsx
@@ -10,6 +10,7 @@ import LayerControlsPanel from "./LayerControlsPanel";
 import DetailPanel from "./DetailPanel";
 import EditingToolsPanel from "./EditingToolsPanel";
 import { CheckboxState } from "./GeoSetItem";
+import { createCircleBuffer } from "@/lib/geo/buffer";
 import { ConsultationCommentWithUpvotes } from "@/lib/db/consultations";
 import { Location } from "@/lib/types/onboarding";
 import { useIsMobile } from "@/hooks/use-mobile";
@@ -40,38 +41,6 @@ const GEOSET_COLORS = [
     '#D53F8C', // Pink
     '#4A5568', // Gray
 ];
-
-// Helper function to create a circular polygon buffer around a point
-function createCircleBuffer(center: [number, number], radiusInMeters: number): GeoJSON.Polygon {
-    const earthRadius = 6371000; // Earth's radius in meters
-    const lat = center[1] * Math.PI / 180; // Convert to radians
-    const lng = center[0] * Math.PI / 180;
-
-    const points: [number, number][] = [];
-    const numPoints = 64; // Number of points to create the circle
-
-    for (let i = 0; i < numPoints; i++) {
-        const angle = (i * 360 / numPoints) * Math.PI / 180;
-
-        // Calculate offset in radians
-        const dLat = radiusInMeters * Math.cos(angle) / earthRadius;
-        const dLng = radiusInMeters * Math.sin(angle) / (earthRadius * Math.cos(lat));
-
-        // Convert back to degrees
-        const newLat = (lat + dLat) * 180 / Math.PI;
-        const newLng = (lng + dLng) * 180 / Math.PI;
-
-        points.push([newLng, newLat]);
-    }
-
-    // Close the polygon by adding the first point at the end
-    points.push(points[0]);
-
-    return {
-        type: 'Polygon',
-        coordinates: [points]
-    };
-}
 
 // Helper function to compute derived geometry
 function computeDerivedGeometry(derivedGeometry: DerivedGeometry, allGeoSets: GeoSetData[]): GeoJSON.Geometry | null {
